!!![FEATURE] New time locking mechanism
When the user opens a form the timestamp is stored in the current user
session with a unique id. This id is rendered into the form.

When the user submits the form the timestamp is validated. If the
submission was too quick or takes too long the user will be blocked.

Additionally many deprecated API calls have been replaced with their
new namespaced versions. This breaks compatibility with all TYPO3 version
prior 6.0.